Remove Sys.Reactive dependency & Periodic feature
MatchaBackgroundService first feature was to run a task many times
(periodic), and later it allowed to run it only once if the task
returned false.

To me, it's more important to cover this "only once" functionality,
without confusing APIs that are related only to the "periodic" part,
such as "Interval", or the "Periodic" word in the interface name
"IPeriodicTask", etc.

So in this commit I'm removing the periodic bits from the API, the
periodic functionality, and System.Reactive dependency because it
seemed it was only needed for this.

We may revisit adding this back in the future (but without the nuget
dependency System.Reactive, because it brings too many subdependencies).